requirements.txt
aioredis==1.2.0
alabaster==0.7.12
amqp==2.5.0
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==19.1.0
asgiref==3.1.4
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.7.1
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
billiard==3.6.0.0
black==19.3b0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
cryptography==2.7
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-celery-beat==1.5.0
django-compressor==2.3
django-contrib-comments==1.9.1
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.9
django-haystack==2.8.1
django-markdownx==2.0.28
django-model-utils==3.1.2
django-redis==4.10.0
django-taggit==1.1.0
django-test-plus==1.2.0
django-timezone-field==3.0
djangorestframework==3.9.4
docutils==0.14
elasticsearch==2.4.1
entrypoints==0.3
factory-boy==2.12.0
Faker==1.0.7
flake8==3.7.7
future==0.17.1
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
incremental==17.5.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10.1
kombu==4.6.3
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
msgpack==0.6.1
mypy==0.701
mypy-extensions==0.4.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
packaging==19.0
parso==0.4.0
pefile==2019.4.18
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
prompt-toolkit==2.0.9
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyHamcrest==1.9.0
pylint==2.3.1
pylint-celery==0.3
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyparsing==2.4.0
pytest==4.6.3
pytest-django==3.5.0
pytest-sugar==0.9.2
python-crontab==2.3.6
python-dateutil==2.8.0
python-ptrace==0.9.3
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2019.1
rcssmin==1.0.6
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
rjsmin==1.1.0
six==1.12.0
snowballstemmer==1.2.1
sorl-thumbnail==12.5.0
Sphinx==2.1.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.0
traitlets==4.3.2
Twisted==19.2.1
txaio==18.8.1
typed-ast==1.3.5
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.11.2
zipp==0.5.1
zope.interface==4.6.0
